12/1/2025: Hoàn thành cập nhật dữ liệu thống kê khi đơn đặt hàng được xử lý

+++ E-Commerce/Web/Models/StatisticalChartModel.cs
@@ -8,8 +8,8 @@ namespace Web.Models
         public int Id { get; set; }
         public int Quantity { get; set; }
         public int Sold { get; set; }
-        public int Revenue { get; set; }
-        public int Profit { get; set; }
+        public decimal Revenue { get; set; }
+        public decimal Profit { get; set; }
         public DateTime CreatedDate { get; set; }
     }
 }
